# scratch_swagger_new.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "StreamVault",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/Movie/latest":  {
                                            "get":  {
                                                        "tags":  [
                                                                     "Movie"
                                                                 ],
                                                        "parameters":  [
                                                                           {
                                                                               "name":  "genres",
                                                                               "in":  "query",
                                                                               "schema":  {
                                                                                              "type":  "array",
                                                                                              "items":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                          }
                                                                           }
                                                                       ],
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "OK"
                                                                                  }
                                                                      }
                                                    }
                                        },
                  "/api/Movie/by-genres":  {
                                               "get":  {
                                                           "tags":  [
                                                                        "Movie"
                                                                    ],
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "pageNumber",
                                                                                  "in":  "query",
                                                                                  "schema":  {
                                                                                                 "type":  "integer",
                                                                                                 "format":  "int32",
                                                                                                 "default":  1
                                                                                             }
                                                                              },
                                                                              {
                                                                                  "name":  "pageSize",
                                                                                  "in":  "query",
                                                                                  "schema":  {
                                                                                                 "type":  "integer",
                                                                                                 "format":  "int32",
                                                                                                 "default":  10
                                                                                             }
                                                                              },
                                                                              {
                                                                                  "name":  "genres",
                                                                                  "in":  "query",
                                                                                  "schema":  {
                                                                                                 "type":  "array",
                                                                                                 "items":  {
                                                                                                               "type":  "string"
                                                                                                           }
                                                                                             }
                                                                              }
                                                                          ],
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK"
                                                                                     }
                                                                         }
                                                       }
                                           },
...
                  "/api/Movie/{movieGuid}":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "Movie"
                                                                      ],
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "movieGuid",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "string"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK"
                                                                                       }
                                                                           }
                                                         },
                                                 "put":  {
                                                             "tags":  [
                                                                          "Movie"
                                                                      ],
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "movieGuid",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "string"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK"
                                                                                       }
                                                                           }
                                                         }
                                             },
...
                  "/api/Review/movie/{movieGuid}":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "Review"
                                                                             ],
                                                                    "parameters":  [
                                                                                       {
                                                                                           "name":  "movieGuid",
                                                                                           "in":  "path",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "string"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "OK"
                                                                                              }
                                                                                  }
                                                                }
                                                    },
...
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "apiKey",
                                                              "description":  "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter \u0027Bearer\u0027 [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\"",
                                                              "name":  "Authorization",
                                                              "in":  "header"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ],
...
}